% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{commentary}
\alias{commentary}
\title{Emojidome Commentary Log (XKCD 2131)}
\format{A data frame with 1,263 rows and 5 variables:
\describe{
  \item{round_id}{indentifer for collecting all games in one round}
  \item{game_id}{unique ID for the game}
  \item{end_date}{end date and time of the game}
  \item{comment_id}{sequential per-game comment ID}
  \item{comment}{commentary string}
}}
\source{
\url{https://emojidome.xkcd.com/2131/socket}
}
\usage{
commentary
}
\description{
A dataset containing per-game commentary strings. Commentary is indexed in
ascending order (older comments come first) as far as I can tell.
}
\details{
XKCD comics are released under a Creative Commons BY-NC license.
}
\keyword{datasets}
