% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emojitracker}
\alias{emojitracker}
\title{Emojitracker Data Set}
\format{A data frame with 845 rows and 4 variables:
\describe{
  \item{emoji_id}{emojitracker (UTF-8 derived) indentifer for an emoji}
  \item{emoji_text}{alternate text or description for an emoji}
  \item{emoji}{emoji character}
  \item{score}{usage score as reported by emojitracker}
}}
\source{
\url{http://emojitracker.com}
}
\usage{
emojitracker
}
\description{
A dataset containing scores and metadata from \url{http://emojitracker.com},
downloaded April 2, 2019.
}
\details{
The scores are *very* large integers, stored as \code{numeric} type.
}
\keyword{datasets}
