% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{games}
\alias{games}
\title{Emojidome Game Log (XKCD 2131)}
\format{A data frame with 1,022 rows and 5 variables:
\describe{
  \item{round_id}{indentifer for collecting all games in one round}
  \item{game_id}{unique ID for the game}
  \item{end_date}{end date and time of the game}
  \item{competitor}{a competitor emoji}
  \item{score}{final score for a competitor in the specified game}
}}
\source{
\url{https://emojidome.xkcd.com/2131/socket}
}
\usage{
games
}
\description{
A dataset containing game identifiers, players, and scores.
}
\details{
XKCD comics are released under a Creative Commons BY-NC license.
}
\keyword{datasets}
